fix(gitlab): modify table field name
