+++ plugins/gitlab/tasks/gitlab_commit_collector.go
@@ -51,7 +51,7 @@ func CollectCommits(projectId int, gitlabApiClient *GitlabApiClient) error {
 			}
 			gitlabProjectCommit := &models.GitlabProjectCommit{GitlabProjectId: projectId}
 			for _, gitlabApiCommit := range *gitlabApiResponse {
-				gitlabCommit, err := convertCommit(&gitlabApiCommit, projectId)
+				gitlabCommit, err := ConvertCommit(&gitlabApiCommit)
 				if err != nil {
 					return err
 				}
@@ -103,7 +103,7 @@ func CollectCommits(projectId int, gitlabApiClient *GitlabApiClient) error {
 }
 
 // Convert the API response to our DB model instance
-func convertCommit(commit *GitlabApiCommit, projectId int) (*models.GitlabCommit, error) {
+func ConvertCommit(commit *GitlabApiCommit) (*models.GitlabCommit, error) {
 	gitlabCommit := &models.GitlabCommit{
 		Sha:            commit.GitlabId,
 		Title:          commit.Title,
